libbpf-tools: Fix 64-bit format portability

+++ libbpf-tools/filelife.c
@@ -117,7 +117,7 @@ int handle_event(void *ctx, void *data, size_t data_sz)
 
 void handle_lost_events(void *ctx, int cpu, __u64 lost_cnt)
 {
-	fprintf(stderr, "lost %llu events on CPU #%d\n", lost_cnt, cpu);
+	fprintf(stderr, "lost %llu events on CPU #%d\n", (unsigned long long)lost_cnt, cpu);
 }
 
 int main(int argc, char **argv)
