+++ libbpf-tools/opensnoop.c
@@ -296,7 +296,7 @@ int handle_event(void *ctx, void *data, size_t data_sz)
 
 void handle_lost_events(void *ctx, int cpu, __u64 lost_cnt)
 {
-	fprintf(stderr, "Lost %llu events on CPU #%d!\n", lost_cnt, cpu);
+	fprintf(stderr, "Lost %llu events on CPU #%d!\n", (unsigned long long)lost_cnt, cpu);
 }
 
 int main(int argc, char **argv)
